Developed Food functionality

+++ SnakeGame/Snake.h
@@ -4,18 +4,18 @@
 class snake
 {
 private:
-	sf::Vector2f _direction = {0,1};
+	sf::Vector2i _direction = {0,1};
 	int _snakeLength;
-	sf::Vector2f _bodycoordinates[64];
+	sf::Vector2i _bodycoordinates[64];
 
 public:
 	void initialize();
 	void setDirection();
-	sf::Vector2f getDirection();
+	sf::Vector2i getDirection();
 	void setLength(int i);
 	int getLength();
 	void setCoordinates();
-	sf::Vector2f getCoordinates(int i);
+	sf::Vector2i getCoordinates(int i);
 	void reset();
 	bool deathCondition();
 };
